+++ packages/fiori-elements-writer/src/data/defaults.ts
@@ -5,6 +5,7 @@ import type {
     ALPSettings,
     ALPSettingsV2,
     ALPSettingsV4,
+    FioriApp,
     FioriElementsApp,
     LROPSettings,
     Template,
@@ -21,10 +22,12 @@ import {
 import { getAnnotationV4Libs } from './annotationCustomUi5Libs';
 import { type TemplateOptions } from './templateAttributes';
 import semVer from 'semver';
+import { getFlpId } from '@sap-ux/fiori-generator-shared';
 
 const defaultModelName = 'mainModel'; // UI5 default model name is '' but some floorplans require a named default model
 const defaultVirtualPreviewFile = 'test/flp.html'; // Default virtual preview file name
 const defaultIntent = 'app-preview';
+const defaultNavActionTile = 'tile';
 
 /**
  * Sets defaults for relevant parameters (`flpAppId`, `startFile`, `localStartFile`,  ) when virtual endpoints are used.
@@ -113,6 +116,22 @@ export function getManifestLibs(type: TemplateType, version: OdataVersion, libs?
 }
 
 /**
+ *  Set defaults for missing parameters on the given Fiori/UI5 app instance.
+ *
+ * @param app - fiori app config
+ * @param templateType - template type
+ * @param serviceVersion - odata version for the service
+ */
+function setFioriAppDefaults(app: FioriApp, templateType: TemplateType, serviceVersion: OdataVersion) {
+    // Generate base UI5 project
+    // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
+    app.baseComponent = app.baseComponent || getBaseComponent(templateType, serviceVersion);
+    app.flpAction = app.flpAction || defaultNavActionTile;
+    // create `flpAppId` if not already added
+    app.flpAppId = app.flpAppId || getFlpId(app.id, app.flpAction);
+}
+
+/**
  * Sets defaults for the specified Fiori elements application.
  *
  * @param feApp - Fiori elements application config
@@ -129,9 +148,7 @@ export function setAppDefaults<T>(feApp: FioriElementsApp<T>): FioriElementsApp<
         };
     }
 
-    // Generate base UI5 project
-    // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
-    feApp.app.baseComponent = feApp.app.baseComponent || getBaseComponent(feApp.template.type, feApp.service.version);
+    setFioriAppDefaults(feApp.app, feApp.template.type, feApp.service.version);
 
     const customUi5Libs =
         feApp.service.version === OdataVersion.v4 && feApp.service.metadata