+++ packages/fiori-generator-shared/src/index.ts
@@ -10,3 +10,4 @@ export { getBootstrapResourceUrls } from './ui5/ui5';
 export { getDefaultTargetFolder, isExtensionInstalled } from './vscode-helpers/vscode-helpers';
 export { generateAppGenInfo } from './app-gen-info';
 export { getHostEnvironment } from './environment';
+export { getFlpId, getSemanticObject } from './app-helpers/app-helpers';